StudentRepository first implementation

+++ BDSA2020.Models/StudentRepository.cs
@@ -7,10 +7,6 @@ using Microsoft.EntityFrameworkCore;
 
 namespace BDSA2020.Models
 {
-    public interface IStudentRepository
-    {
-        Task<ICollection<Student>> GetStudentsAsync();
-    }
     public class StudentRepository : IStudentRepository
     {
         private readonly IContext context;
@@ -60,7 +56,35 @@ namespace BDSA2020.Models
         {
             var student = await context.Students.FindAsync(id);
 
+            if (student == null) 
+            {
+                throw new ArgumentException($"Could not find student with id {id}");
+            }
+
             context.Students.Remove(student);
+
+            await context.SaveChangesAsync();
+
+            return true;
+        }
+
+        public async Task<bool> UpdateStudentAsync(Student student)
+        {
+            var entity = await context.Students.FindAsync(student.Id);
+
+            if (entity == null)
+            {
+                throw new ArgumentException("Student does not exist");
+            }
+
+            entity.Agreement = student.Agreement;
+            entity.Degree = student.Degree;
+            entity.Keywords = student.Keywords;
+            entity.Location = student.Location;
+            entity.MaxWorkingHours = student.MaxWorkingHours;
+            entity.MinWorkingHours = student.MinWorkingHours;
+            entity.MinSalary = student.MinSalary;
+            entity.PlacementDescriptions = student.PlacementDescriptions;
 
             await context.SaveChangesAsync();
